Count CAAP budget unit by eligible persons instead of subtracting ineligible
Replace ca_sf_caap_ineligible_person with a positive ca_sf_caap_eligible_person
(not an SSI recipient and has a qualified immigration status). budget_unit_size
now adds eligible persons directly (dropping the spm_unit_size dependency), and
countable_income keeps eligible persons' income. Behavior is identical
(count(eligible) == spm_unit_size - count(ineligible)).
